+++ extension.js
@@ -1,7 +1,8 @@
-// The module 'vscode' contains the VS Code extensibility API
-// Import the module and reference it with the alias vscode in your code below
 const vscode = require('vscode');
 const which = require('which');
+const { dirname } = require('path');
+const fs = require('fs');
+const { readConfig } = require('./configReader.js');
 
 function detectOS() {
 	switch (process.platform) {
@@ -12,9 +13,6 @@ function detectOS() {
 }
 
 
-
-// this method is called when your extension is activated
-// your extension is activated the very first time the command is executed
 
 /**
  * @param {vscode.ExtensionContext} context
@@ -63,56 +61,66 @@ function activate(context) {
 	var newTerminal = vscode.workspace.getConfiguration('aspLanguage').get("terminalMode");
 	var turnMessagesOff = vscode.workspace.getConfiguration('aspLanguage').get("turnMessagesOff");
 	var usePathClingo = vscode.workspace.getConfiguration('aspLanguage').get("usePathClingo");
-
+	var additionalArgs = vscode.workspace.getConfiguration('aspLanguage').get("additionalArgs");
+	var setConfig = vscode.workspace.getConfiguration('aspLanguage').get("setConfig");
 	var path;
-	var n = (vscode.window).terminals.length;
+	var terminalCount = (vscode.window).terminals.length;
 
-	getNewPath()
+	getNewPath();
 
 	//double if to prevent unnecessary info messages
 	if (usePathClingo) {
 		usePath()
 	}
 
-	let disposable = vscode.commands.registerCommand('answer-set-programming-language-support.runinterminalall', function () {
-		if (newTerminal == true) {
-			n = (vscode.window).terminals.length + 1;
+	function createTerminal() {
+		if (newTerminal || terminalCount === 0) {
+			terminalCount = (vscode.window).terminals.length + 1;
 			terminal = vscode.window.createTerminal("ASP Terminal");
-
 		}
-		if ((vscode.window).terminals.length === 0) {
-			n = (vscode.window).terminals.length + 1;
-			terminal = vscode.window.createTerminal("ASP Terminal");
+	}
+	
 
-		}
-		//vscode.window.showInformationMessage("Running in Terminal " + n + " ...");
+	const computeAllSetsCommand = vscode.commands.registerCommand('answer-set-programming-language-support.runinterminalall', function () {
+		createTerminal();
+
 		terminal.show();
-		terminal.sendText(path + " " + vscode.window.activeTextEditor.document.fileName + " 0");
+		terminal.sendText(`${path} ${vscode.window.activeTextEditor.document.fileName} 0`);
 	});
 
-	let disposable2 = vscode.commands.registerCommand('answer-set-programming-language-support.runinterminalsingle', function () {
-		if (newTerminal == true) {
-			n = (vscode.window).terminals.length + 1;
+	const computeSingleSetCommand = vscode.commands.registerCommand('answer-set-programming-language-support.runinterminalsingle', function () {
+		createTerminal();
 
-			terminal = vscode.window.createTerminal("ASP Terminal");
-
-		}
-		if ((vscode.window).terminals.length === 0) {
-			n = (vscode.window).terminals.length + 1;
-			terminal = vscode.window.createTerminal("ASP Terminal");
-			
-		}
-		//vscode.window.showInformationMessage("Running in Terminal " + n + " ...");
 		terminal.show();
-		terminal.sendText(path + " " + vscode.window.activeTextEditor.document.fileName);
+		terminal.sendText(`${path} ${vscode.window.activeTextEditor.document.fileName}`);
 	});
 
-	context.subscriptions.push(disposable);
-	context.subscriptions.push(disposable2);
+	const computeConfigCommand = vscode.commands.registerCommand('answer-set-programming-language-support.runinterminalconfig', function () {
+		createTerminal();
+
+		additionalArgs = readConfig(setConfig, turnMessagesOff, context.asAbsolutePath(""));
+
+		terminal.show();
+		terminal.sendText(`${path} ${vscode.window.activeTextEditor.document.fileName} ${additionalArgs}`);
+	});
+
+	const initClingoConfig = vscode.commands.registerCommand('answer-set-programming-language-support.initClingoConfig', function () {
+		createTerminal();
+
+		const sampleConfig = fs.readFileSync(`${context.asAbsolutePath("")}\\sampleConfig.json`);
+		fs.writeFileSync(`${dirname(vscode.window.activeTextEditor.document.fileName)}\\config.json`, sampleConfig);
+		
+		vscode.workspace.getConfiguration('aspLanguage').update("setConfig", "config.json");
+	});
+
+	context.subscriptions.push(computeAllSetsCommand);
+	context.subscriptions.push(computeSingleSetCommand);
+	context.subscriptions.push(computeConfigCommand);
+	context.subscriptions.push(initClingoConfig);
 
 	//Listeners for Configuration Options 
 	vscode.workspace.onDidChangeConfiguration(event => {
-        let affected = event.affectsConfiguration("aspLanguage.selectOperatingSystem");
+        const affected = event.affectsConfiguration("aspLanguage.selectOperatingSystem");
         if (affected) {
 			os = vscode.workspace.getConfiguration('aspLanguage').get("selectOperatingSystem");
 			getNewPath()
@@ -120,29 +128,36 @@ function activate(context) {
 	})
 
 	vscode.workspace.onDidChangeConfiguration(event => {
-        let affected = event.affectsConfiguration("aspLanguage.terminalMode");
+        const affected = event.affectsConfiguration("aspLanguage.terminalMode");
         if (affected) {
 			newTerminal = vscode.workspace.getConfiguration('aspLanguage').get("terminalMode");
         }
 	})
 
 	vscode.workspace.onDidChangeConfiguration(event => {
-        let affected = event.affectsConfiguration("aspLanguage.turnMessagesOff");
+        const affected = event.affectsConfiguration("aspLanguage.turnMessagesOff");
         if (affected) {
 			turnMessagesOff = vscode.workspace.getConfiguration('aspLanguage').get("turnMessagesOff");
         }
 	})
 	
 	vscode.workspace.onDidChangeConfiguration(event => {
-        let affected = event.affectsConfiguration("aspLanguage.usePathClingo");
+        const affected = event.affectsConfiguration("aspLanguage.usePathClingo");
         if (affected) {
 			usePathClingo = vscode.workspace.getConfiguration('aspLanguage').get("usePathClingo");
 			usePath()
         }
+	})
+	
+	vscode.workspace.onDidChangeConfiguration(event => {
+        const affected = event.affectsConfiguration("aspLanguage.setConfig");
+        if (affected) {
+			setConfig = vscode.workspace.getConfiguration('aspLanguage').get("setConfig");
+			
+        }
     })
 
 }
-exports.activate = activate;
 
 // this method is called when your extension is deactivated
 function deactivate() {